pe code cave method is the same as elfs, so making a common function for these
